+++ client/src/components/Success.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from "react";
-import { useQuery, useMutation } from "@apollo/react-hooks";
+import { useParams } from "react-router-dom";
+import { useMutation } from "@apollo/react-hooks";
 import { COMPLETE_JOB } from "../utils/mutation";
 import { Button } from "react-bootstrap";
 
@@ -8,17 +9,16 @@ const Success = () => {
   var timeRemaining = seconds;
 
   const [completeJob] = useMutation(COMPLETE_JOB);
-  const queryString = window.location.search;
-  const urlParams = new URLSearchParams(queryString);
-  const job_Id = urlParams.get("id");
+  const { job_Id } = useParams();
 
   completeJob({
     variables: { _id: job_Id },
   });
 
+  // Timer to reroute
   const countDown = () => {
     timeRemaining--;
-    if (timeRemaining > 0 || timeRemaining === 1) {
+    if (timeRemaining > 0) {
       setSeconds(timeRemaining);
     } else {
       window.location.assign("/profile");
